feat: add method that reads buffer into vector (#127)
* feat: add method that reads buffer into vector

* fix: potential offset overflow


* fix: fix compilation

* chore: fix clippy warnings

* chore: fix more clippy warnings

* chore: finished fixing clippy warnings

---------

+++ snippets/i64_shr_s.rs
@@ -1,7 +1,7 @@
 #[inline(always)]
 pub(crate) fn i64_shr_s_impl(a_lo: u32, a_hi: u32, b_lo: u32, _b_hi: u32) -> (u32, u32) {
     // WASM uses only the low 6 bits of the shift count
-    let shamt = (b_lo & 0x3F) as u32;
+    let shamt = b_lo & 0x3F;
 
     match shamt {
         0 => (a_lo, a_hi),
